database-schema-to-pr-agent/pyproject.toml
[project]
name = "schema-to-pr-agent"
version = "0.1.0"
description = "AI Agent that updates database schemas and creates GitHub PRs with matching Python model classes"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "agno>=1.7.0",
    "anthropic>=0.55.0",
    "python-dotenv>=1.1.1",
    "groq>=0.20.0",
    "mcp>=1.10.0",
    "openai>=1.70.0",
    "streamlit>=1.44.0",
    "pydantic>=2.11.7",
    "gitpython>=3.1.40",
    "sqlalchemy>=2.0.41",
]
